SafeBet-Analyst/requirements.txt
playwright==1.40.0
streamlit==1.29.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
openai==1.3.7
python-dotenv==1.0.0